refactor: extract auth options from pages router
