+++ app/(auth)/auth/confirm-email-change/[token]/utils.ts
@@ -1,4 +1,4 @@
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/lib/auth/auth-options";
 import { getServerSession } from "next-auth";
 
 export const getSession = async () => {
