+++ app/(ee)/api/ai/chat/route.ts
@@ -2,7 +2,7 @@ import { NextRequest, NextResponse } from "next/server";
 
 import { createChat } from "@/ee/features/ai/lib/chat/create-chat";
 import { createChatSchema } from "@/ee/features/ai/schemas/chat";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/lib/auth/auth-options";
 import { getServerSession } from "next-auth";
 import { z } from "zod";
