+++ app/(ee)/api/workflows/[workflowId]/route.ts
@@ -4,7 +4,7 @@ import {
   UpdateWorkflowRequestSchema,
   formatZodError,
 } from "@/ee/features/workflows/lib/validation";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/lib/auth/auth-options";
 import { customAlphabet } from "nanoid";
 import { getServerSession } from "next-auth";
 import { z } from "zod";
