+++ app/(ee)/api/workflows/[workflowId]/steps/[stepId]/route.ts
@@ -6,7 +6,7 @@ import {
   validateActions,
   validateConditions,
 } from "@/ee/features/workflows/lib/validation";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/lib/auth/auth-options";
 import { getServerSession } from "next-auth";
 import { z } from "zod";
