+++ app/(ee)/api/workflows/[workflowId]/steps/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/lib/auth/auth-options";
 import { z } from "zod";
 import prisma from "@/lib/prisma";
 import { CustomUser } from "@/lib/types";
