+++ app/(ee)/api/workflows/route.ts
@@ -4,7 +4,7 @@ import {
   CreateWorkflowRequestSchema,
   formatZodError,
 } from "@/ee/features/workflows/lib/validation";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/lib/auth/auth-options";
 import { getServerSession } from "next-auth";
 import { z } from "zod";
