+++ app/api/integrations/slack/oauth/callback/route.ts
@@ -1,7 +1,7 @@
 import { redirect } from "next/navigation";
 import { NextResponse } from "next/server";
 
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/lib/auth/auth-options";
 import { Team } from "@prisma/client";
 import { getServerSession } from "next-auth";
 import z from "zod";
