+++ ee/features/billing/cancellation/api/retention-route.ts
@@ -3,7 +3,7 @@ import { NextApiRequest, NextApiResponse } from "next";
 import { stripeInstance } from "@/ee/stripe";
 import { getCouponFromPlan } from "@/ee/stripe/functions/get-coupon-from-plan";
 import { isOldAccount } from "@/ee/stripe/utils";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/lib/auth/auth-options";
 import { waitUntil } from "@vercel/functions";
 import { getServerSession } from "next-auth/next";
