+++ ee/features/dataroom-invitations/api/group-invite.ts
@@ -5,7 +5,7 @@ import {
   MAX_INVITATION_EMAILS_PER_REQUEST,
   SendGroupInvitationSchema,
 } from "@/ee/features/dataroom-invitations/lib/schema/dataroom-invitations";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/lib/auth/auth-options";
 import { LinkAudienceType, LinkType } from "@prisma/client";
 import { getServerSession } from "next-auth/next";
